+++ lib/td_dd/data_structures/duplicate_remover.ex
@@ -20,11 +20,12 @@ defmodule TdDd.DataStructures.DuplicateRemover do
 
   require Logger
 
-  def start_link(arg) do
-    Task.start_link(__MODULE__, :run, [arg])
+  def start_link(_arg) do
+    env = Application.get_env(:td_dd, :env)
+    Task.start_link(__MODULE__, :run, [env])
   end
 
-  def run(_options), do: :ok
+  def run(:test), do: :ok
 
   def run(_options) do
     query = """
@@ -37,7 +38,6 @@ defmodule TdDd.DataStructures.DuplicateRemover do
 
     ids = Repo |> SQL.query!(query) |> Map.get(:rows) |> Enum.flat_map(& &1)
 
-    # credo:disable-for-lines:2 Credo.Check.Refactor.PipeChainStart
     from(v in DataStructureVersion, where: v.id in ^ids, select: v.data_structure_id)
     |> Repo.delete_all()
     |> post_process()
@@ -50,7 +50,6 @@ defmodule TdDd.DataStructures.DuplicateRemover do
   defp post_process({count, data_structure_ids}) do
     Logger.info("Deleted #{count} duplicate data structure versions")
 
-    # credo:disable-for-lines:2 Credo.Check.Refactor.PipeChainStart
     from(ds in DataStructure, where: ds.id in ^data_structure_ids)
     |> Repo.all()
     |> IndexWorker.reindex()